ai/app/requirements.txt
fastapi
uvicorn
pydantic-settings>=2.0.3
chromadb
sentence-transformers>=2.2.2
llama-cpp-python